requirements.txt
diffusers
moviepy
transformers
soundfile
accelerate
opencv-python
pydub
scipy
pyttsx3
python-dotenv
plyvel
rich